web/requirements.txt
Flask
flask_wtf
wtforms
flask-sqlalchemy
requests
flask_migrate
argon2-cffi
flask_sqlalchemy
python-dotenv
flask-cors
gunicorn
psycopg-binary

openai
langchain
langchain-community
